+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   subqueries occurring in a single-value context and outlined in Chapter 9 of the PartiQL specification. 
   This is backward incompatible with the prior behavior (which left the computed collection as is), 
   but brings it in conformance with the specification.
+- Added `partiql-plan` package which contains experimental PartiQL Plan data structures.
 
 ### Changed
 - Deprecates the project level opt-in annotation `PartiQLExperimental` and split it into feature level. `ExperimentalPartiQLCompilerPipeline` and `ExperimentalWindowFunctions`.